notification/django/requirements.txt
djangorestframework~=3.14.0
pika~=1.3.1